webpage crawler
This web crawler is currently almost functional. It is able to pull page
links, and then use spatula to generate the text files. However not all
recipes appear to be producing valid results, and there is an
inconsistence between one page of the results and the next. These bugs
need to be fixed, and also a more sophisticated error detection and
handling system needs to be implemented