Bug fixed and new bug founds
-Fixed bug created when a recipe doesnt have an item. Now the module
will check to see if an object is the correct type before it attempts to
write the recipe .txt file for the wrap function. This serves as a
primitive error checking mechanism
-There is still a bug though with the Carrot Cake III on page 2, for
some reason the ingredients are not getting picked up. Check to make
sure that something isnt going on with the error checking